refactor: update imports to use shadcn directory in learning and portal pages
